feat: run the jwtVerify directly in routes and refactor
