+++ packages/mosip-api/src/index.ts
@@ -1,4 +1,4 @@
-import Fastify from "fastify";
+import Fastify, { FastifyInstance } from "fastify";
 import {
   serializerCompiler,
   validatorCompiler,
@@ -20,7 +20,7 @@ import formbody from "@fastify/formbody";
 import { reviewEventHandler } from "./routes/event-review";
 import cors from "@fastify/cors";
 import jwt from "@fastify/jwt";
-import { getPublicKey, withAuthentication } from "./jwt-auth";
+import { getPublicKey } from "./opencrvs-api";
 
 const envToLogger = {
   development: {
@@ -33,31 +33,12 @@ const envToLogger = {
   },
   production: true,
 };
-const app = Fastify({
-  logger: envToLogger[env.isProd ? "production" : "development"],
-});
 
-app.setValidatorCompiler(validatorCompiler);
-app.setSerializerCompiler(serializerCompiler);
-app.register(formbody);
-app.register(cors, {
-  origin: [env.CLIENT_APP_URL],
-  methods: ["GET", "POST", "OPTIONS"],
-  allowedHeaders: ["Content-Type", "Authorization"],
-});
-
-openapi.register(app);
-
-app.setErrorHandler((error, request, reply) => {
-  request.log.error(error);
-  reply.status(500).send({ error: "An unexpected error occurred" });
-});
-
-app.after(() => {
+const initRoutes = (app: FastifyInstance) => {
   app.withTypeProvider<ZodTypeProvider>().route({
     url: "/events/registration",
     method: "POST",
-    handler: withAuthentication(registrationEventHandler),
+    handler: registrationEventHandler,
     schema: {
       body: opencrvsRecordSchema,
     },
@@ -65,7 +46,7 @@ app.after(() => {
   app.withTypeProvider<ZodTypeProvider>().route({
     url: "/events/review",
     method: "POST",
-    handler: withAuthentication(reviewEventHandler),
+    handler: reviewEventHandler,
     schema: {
       body: opencrvsRecordSchema,
     },
@@ -73,7 +54,7 @@ app.after(() => {
   app.withTypeProvider<ZodTypeProvider>().route({
     url: "/webhooks/mosip",
     method: "POST",
-    handler: withAuthentication(mosipHandler),
+    handler: mosipHandler,
     schema: {
       body: mosipNidSchema,
     },
@@ -88,13 +69,47 @@ app.after(() => {
       querystring: OIDPQuerySchema,
     },
   });
-});
+};
 
-async function run() {
+export const buildFastify = async () => {
+  const app = Fastify({
+    logger: envToLogger[env.isProd ? "production" : "development"],
+  });
+
+  app.setValidatorCompiler(validatorCompiler);
+  app.setSerializerCompiler(serializerCompiler);
+  app.register(formbody);
+  app.register(cors, {
+    origin: [env.CLIENT_APP_URL],
+    methods: ["GET", "POST", "OPTIONS"],
+    allowedHeaders: ["Content-Type", "Authorization"],
+  });
+
+  openapi.register(app);
+
+  app.setErrorHandler((error, request, reply) => {
+    request.log.error(error);
+    reply.status(500).send({ error: "An unexpected error occurred" });
+  });
+
   app.register(jwt, {
     secret: { public: await getPublicKey() },
     verify: { algorithms: ["RS256"] },
   });
+
+  app.after(() => initRoutes(app));
+
+  return app;
+};
+
+async function run() {
+  // Only run the daemon if it's executed directly - as in `tsx index.ts` for example
+  if (require.main !== module) {
+    return;
+  }
+
+  const app = await buildFastify();
+
   await app.ready();
   await app.listen({
     port: env.PORT,
