+++ packages/mosip-api/src/opencrvs-api.ts
@@ -7,6 +7,23 @@ export class OpenCRVSError extends Error {
   }
 }
 
+/** Fetches the public key from OpenCRVS to be able to verify JWTs */
+export const getPublicKey = async (): Promise<string> => {
+  try {
+    const response = await fetch(env.OPENCRVS_PUBLIC_KEY_URL);
+    return response.text();
+  } catch (error) {
+    console.error(
+      `🔑  Failed to fetch public key from Core. Make sure Core is running, and you are able to connect to ${env.OPENCRVS_PUBLIC_KEY_URL}`,
+    );
+    if (env.isProd) {
+      throw error;
+    }
+    await new Promise((resolve) => setTimeout(resolve, 3000));
+    return getPublicKey();
+  }
+};
+
 /** Communicates with opencrvs-core's GraphQL gateway */
 const post = async ({
   query,
