+++ packages/mosip-api/src/routes/event-registration.ts
@@ -29,19 +29,16 @@ export const registrationEventHandler = async (
   request: OpenCRVSRequest,
   reply: FastifyReply,
 ) => {
+  try {
+    await request.jwtVerify();
+  } catch {
+    return reply.status(401).send({ error: "Unauthorized" });
+  }
+
   const trackingId = getTrackingId(request.body);
   const { id: eventId } = getComposition(request.body);
 
-  if (!request.headers.authorization) {
-    return reply.code(401).send({ error: "Authorization header is missing" });
-  }
-
-  const token = request.headers.authorization.split(" ")[1];
-  if (!token) {
-    return reply
-      .code(401)
-      .send({ error: "Token is missing in Authorization header" });
-  }
+  const token = request.headers.authorization!.split(" ")[1];
 
   // We can trust the token as when `confirmRegistration` or `rejectRegistration` are called, the token is verified by OpenCRVS
   // This server should also only be deployed in the local network so no external calls can be made.
