+++ packages/mosip-api/src/routes/event-review.ts
@@ -69,16 +69,13 @@ export const reviewEventHandler = async (
   request: OpenCRVSRequest,
   reply: FastifyReply,
 ) => {
-  if (!request.headers.authorization) {
-    return reply.code(401).send({ error: "Authorization header is missing" });
+  try {
+    await request.jwtVerify();
+  } catch {
+    return reply.status(401).send({ error: "Unauthorized" });
   }
 
-  const token = request.headers.authorization.split(" ")[1];
-  if (!token) {
-    return reply
-      .code(401)
-      .send({ error: "Token is missing in Authorization header" });
-  }
+  const token = request.headers.authorization!.split(" ")[1];
 
   const informantType = getInformantType(request.body);
   const composition = getComposition(request.body);
